Added in the links.
I added in some links to further information on the topics covered. I chose to use reference style links since the content is already text heavy. I also added in/changed some of the body text to make the links I chose to include fit better with the overall document.